move inputlistener into scene2d/events, add changelog, readme, tests
