+++ actors/src/test/kotlin/kts/actors/eventsTest.kt
@@ -2,6 +2,7 @@ package kts.actors
 
 import com.badlogic.gdx.Input.Keys
 import com.badlogic.gdx.scenes.scene2d.Actor
+import com.badlogic.gdx.scenes.scene2d.Event
 import com.badlogic.gdx.scenes.scene2d.InputEvent
 import com.badlogic.gdx.scenes.scene2d.InputEvent.Type.*
 import com.badlogic.gdx.scenes.scene2d.utils.ChangeListener.ChangeEvent
@@ -124,5 +125,19 @@ class EventsTest {
     event.isFocused = true
     actor.fire(event)
     assertTrue(focused)
+  }
+
+  class MyInputListenerTest : KtxInputListener() {
+    override fun enter(event: InputEvent, x: Float, y: Float, pointer: Int, fromActor: Actor?) = Unit
+    override fun exit(event: InputEvent, x: Float, y: Float, pointer: Int, toActor: Actor?) = Unit
+    override fun keyDown(event: InputEvent, keycode: Int) = false
+    override fun keyTyped(event: InputEvent, character: Char) = false
+    override fun keyUp(event: InputEvent, keycode: Int) = false
+    override fun mouseMoved(event: InputEvent, x: Float, y: Float) = false
+    override fun scrolled(event: InputEvent, x: Float, y: Float, amount: Int) = false
+    override fun touchDown(event: InputEvent, x: Float, y: Float, pointer: Int, button: Int) = false
+    override fun touchDragged(event: InputEvent, x: Float, y: Float, pointer: Int) = Unit
+    override fun touchUp(event: InputEvent, x: Float, y: Float, pointer: Int, button: Int) = Unit
+    override fun handle(event: Event) = false
   }
 }
