+++ app/src/test/kotlin/ktx/app/applicationTest.kt
@@ -122,4 +122,24 @@ class KotlinApplicationTest {
       renderedTimes++
     }
   }
+
+  class MyApplicationListener : KtxApplicationListener {
+    override fun resize(width: Int, height: Int) = Unit
+    override fun create() = Unit
+    override fun render() = Unit
+    override fun resume() = Unit
+    override fun dispose() = Unit
+    override fun pause() = Unit
+  }
+
+  class MyInputProcessor : KtxInputProcessor {
+    override fun touchUp(screenX: Int, screenY: Int, pointer: Int, button: Int) = false
+    override fun keyDown(keycode: Int) = false
+    override fun keyTyped(character: Char) = false
+    override fun keyUp(keycode: Int) = false
+    override fun mouseMoved(screenX: Int, screenY: Int) = false
+    override fun scrolled(amount: Int) = false
+    override fun touchDown(screenX: Int, screenY: Int, pointer: Int, button: Int) = false
+    override fun touchDragged(screenX: Int, screenY: Int, pointer: Int) = false
+  }
 }
